refactor: separate generation/upload of shared objects into two tasks
